table joueur et main

+++ joueur.h
@@ -5,8 +5,8 @@ typedef struct joueur{
     int num;
     char nom[25];
     int en_jeu;
-    int capital;
-    int mise;
+    float capital;
+    float mise;
     struct joueur * suivant;
 } JOUEUR;
